color mood selecting and saving, color saved as comp 1 float at the moment need to work out if thats enough to fetch the correct value

+++ app/src/main/java/com/jbrightman/thedayto/feature_thedayto/domain/model/mood_color/MoodColor.kt
@@ -11,7 +11,7 @@ import com.jbrightman.thedayto.ui.theme.Violet
 @Entity
 data class MoodColor(
     val mood: String,
-    val color: Int,
+    val color: Float,
     val dateStamp: Long,
     @PrimaryKey val id: Int? = null
 ) {
